--- wellforce-backend/wellforce-activity.sql
@@ -1,1 +1,0 @@
-INSERT INTO activity
